# feat: Improve symbols compaction