--- a/after.txt
+++ b/after.txt
@@ -1,0 +1,18 @@
+level=debug msg="cut row group segment" path=/var/folders/6r/0_fmjlmd55b4gtg5fc9pb31c0000gn/T/Test_ValidateBlock1129450014/001/head/01HHYF1R1YXZVTQM9QYD4A6PCF/profiles.0.parquet numProfiles=1
+level=info msg="head successfully written to block" block_path=/var/folders/6r/0_fmjlmd55b4gtg5fc9pb31c0000gn/T/Test_ValidateBlock1129450014/001/local/01HHYF1R1YXZVTQM9QYD4A6PCF
+level=debug msg="cut row group segment" path=/var/folders/6r/0_fmjlmd55b4gtg5fc9pb31c0000gn/T/Test_ValidateBlock855275030/001/head/01HHYF237DJ34Y52A7A25FXWMG/profiles.0.parquet numProfiles=1
+level=info msg="head successfully written to block" block_path=/var/folders/6r/0_fmjlmd55b4gtg5fc9pb31c0000gn/T/Test_ValidateBlock855275030/001/local/01HHYF237DJ34Y52A7A25FXWMG
+level=debug msg="cut row group segment" path=/var/folders/6r/0_fmjlmd55b4gtg5fc9pb31c0000gn/T/Test_ValidateBlock2097283185/001/head/01HHYF2DGYQDNSZQH7JMTKJC2P/profiles.0.parquet numProfiles=1
+level=info msg="head successfully written to block" block_path=/var/folders/6r/0_fmjlmd55b4gtg5fc9pb31c0000gn/T/Test_ValidateBlock2097283185/001/local/01HHYF2DGYQDNSZQH7JMTKJC2P
+level=debug msg="cut row group segment" path=/var/folders/6r/0_fmjlmd55b4gtg5fc9pb31c0000gn/T/Test_ValidateBlock4293766086/001/head/01HHYF2REN596607AA0AN3T7Z6/profiles.0.parquet numProfiles=1
+level=info msg="head successfully written to block" block_path=/var/folders/6r/0_fmjlmd55b4gtg5fc9pb31c0000gn/T/Test_ValidateBlock4293766086/001/local/01HHYF2REN596607AA0AN3T7Z6
+goos: darwin
+goarch: amd64
+pkg: github.com/grafana/pyroscope/pkg/phlaredb
+cpu: Intel(R) Core(TM) i9-9900K CPU @ 3.60GHz
+Benchmark_CompactSplit-16    	       1	6956534524 ns/op	3277083440 B/op	 2184742 allocs/op
+Benchmark_CompactSplit-16    	       1	6882364965 ns/op	3277056048 B/op	 2184722 allocs/op
+Benchmark_CompactSplit-16    	       1	7674093541 ns/op	3277476176 B/op	 2184750 allocs/op
+Benchmark_CompactSplit-16    	       1	6434207180 ns/op	3276393352 B/op	 2184503 allocs/op
+PASS
+ok  	github.com/grafana/pyroscope/pkg/phlaredb	108.114s