--- a/pkg/phlaredb/compact.go
+++ b/pkg/phlaredb/compact.go
@@ -76,10 +76,12 @@ func CompactWithSplitting(ctx context.Context, src []BlockReader, splitCount, st
 		srcMetas[i] = b.Meta()
 	}
 
+	symbolsCompactor := newSymbolsCompactor(dst)
+
 	outMeta := compactMetas(srcMetas...)
 	for _, stage := range splitStages(len(writers), int(stageSize)) {
 		for _, idx := range stage {
-			if writers[idx], err = createBlockWriter(dst, outMeta, splitCount, idx); err != nil {
+			if writers[idx], err = createBlockWriter(dst, outMeta, splitCount, idx, symbolsCompactor.Rewriter); err != nil {
 				return nil, fmt.Errorf("create block writer: %w", err)
 			}
 		}
@@ -114,7 +116,7 @@ func splitStages(n, s int) (stages [][]int) {
 	return stages
 }
 
-func createBlockWriter(dst string, m block.Meta, splitCount uint64, shard int) (*blockWriter, error) {
+func createBlockWriter(dst string, m block.Meta, splitCount uint64, shard int, rewriterFn SymbolsRewriterFn) (*blockWriter, error) {
 	meta := m.Clone()
 	meta.ULID = ulid.MustNew(meta.ULID.Time(), rand.Reader)
 	if splitCount > 1 {
@@ -123,7 +125,7 @@ func createBlockWriter(dst string, m block.Meta, splitCount uint64, shard int) (
 		}
 		meta.Labels[sharding.CompactorShardIDLabel] = sharding.FormatShardIDLabelValue(uint64(shard), splitCount)
 	}
-	return newBlockWriter(dst, meta)
+	return newBlockWriter(dst, meta, rewriterFn)
 }
 
 func compact(ctx context.Context, writers []*blockWriter, readers []BlockReader, splitBy SplitByFunc, splitCount uint64) ([]block.Meta, error) {
@@ -185,14 +187,21 @@ var SplitByStacktracePartition = func(r profileRow, shardsCount uint64) uint64 {
 
 type blockWriter struct {
 	indexRewriter   *indexRewriter
-	symbolsRewriter *symbolsRewriter
+	symbolsRewriter SymbolsRewriter
 	profilesWriter  *profilesWriter
 	path            string
 	meta            *block.Meta
 	totalProfiles   uint64
 }
 
-func newBlockWriter(dst string, meta *block.Meta) (*blockWriter, error) {
+type SymbolsRewriter interface {
+	ReWriteRow(profile profileRow) error
+	Close() (uint64, error)
+}
+
+type SymbolsRewriterFn func(blockPath string) SymbolsRewriter
+
+func newBlockWriter(dst string, meta *block.Meta, rewriterFn SymbolsRewriterFn) (*blockWriter, error) {
 	blockPath := filepath.Join(dst, meta.ULID.String())
 
 	if err := os.MkdirAll(blockPath, 0o777); err != nil {
@@ -206,7 +215,7 @@ func newBlockWriter(dst string, meta *block.Meta) (*blockWriter, error) {
 
 	return &blockWriter{
 		indexRewriter:   newIndexRewriter(blockPath),
-		symbolsRewriter: newSymbolsRewriter(blockPath),
+		symbolsRewriter: rewriterFn(blockPath),
 		profilesWriter:  profileWriter,
 		path:            blockPath,
 		meta:            meta,
@@ -234,7 +243,8 @@ func (bw *blockWriter) Close(ctx context.Context) error {
 	if err := bw.indexRewriter.Close(ctx); err != nil {
 		return err
 	}
-	if err := bw.symbolsRewriter.Close(); err != nil {
+	numSamples, err := bw.symbolsRewriter.Close()
+	if err != nil {
 		return err
 	}
 	if err := bw.profilesWriter.Close(); err != nil {
@@ -247,7 +257,7 @@ func (bw *blockWriter) Close(ctx context.Context) error {
 	bw.meta.Files = metaFiles
 	bw.meta.Stats.NumProfiles = bw.totalProfiles
 	bw.meta.Stats.NumSeries = bw.indexRewriter.NumSeries()
-	bw.meta.Stats.NumSamples = bw.symbolsRewriter.NumSamples()
+	bw.meta.Stats.NumSamples = numSamples
 	bw.meta.Compaction.Deletable = bw.totalProfiles == 0
 	if _, err := bw.meta.WriteToFile(util.Logger, bw.path); err != nil {
 		return err
@@ -642,29 +652,62 @@ func (it *dedupeProfileRowIterator) Next() bool {
 	}
 }
 
-type symbolsRewriter struct {
+type symbolsCompactor struct {
 	rewriters   map[BlockReader]*symdb.Rewriter
 	w           *symdb.SymDB
 	stacktraces []uint32
 
-	numSamples uint64
+	dst     string
+	flushed bool
 }
 
-func newSymbolsRewriter(path string) *symbolsRewriter {
-	return &symbolsRewriter{
+func newSymbolsCompactor(path string) *symbolsCompactor {
+	dst := filepath.Join(path, symdb.DefaultDirName)
+	return &symbolsCompactor{
 		w: symdb.NewSymDB(symdb.DefaultConfig().
-			WithDirectory(filepath.Join(path, symdb.DefaultDirName)).
+			WithDirectory(dst).
 			WithParquetConfig(symdb.ParquetConfig{
 				MaxBufferRowCount: defaultParquetConfig.MaxBufferRowCount,
 			})),
+		dst:       dst,
 		rewriters: make(map[BlockReader]*symdb.Rewriter),
 	}
+}
+
+func (s *symbolsCompactor) Rewriter(dst string) SymbolsRewriter {
+	return &symbolsRewriter{
+		symbolsCompactor: s,
+		dst:              dst,
+	}
+}
+
+type symbolsRewriter struct {
+	*symbolsCompactor
+
+	numSamples uint64
+	dst        string
 }
 
 func (s *symbolsRewriter) NumSamples() uint64 { return s.numSamples }
 
 func (s *symbolsRewriter) ReWriteRow(profile profileRow) error {
-	var err error
+	total, err := s.symbolsCompactor.ReWriteRow(profile)
+	s.numSamples += total
+	return err
+}
+
+func (s *symbolsRewriter) Close() (uint64, error) {
+	if err := s.symbolsCompactor.Close(); err != nil {
+		return 0, err
+	}
+	return s.numSamples, util.CopyDir(s.symbolsCompactor.dst, filepath.Join(s.dst, symdb.DefaultDirName))
+}
+
+func (s *symbolsCompactor) ReWriteRow(profile profileRow) (uint64, error) {
+	var (
+		err              error
+		rewrittenSamples uint64
+	)
 	profile.row.ForStacktraceIDsValues(func(values []parquet.Value) {
 		s.loadStacktracesID(values)
 		r, ok := s.rewriters[profile.blockReader]
@@ -675,23 +718,30 @@ func (s *symbolsRewriter) ReWriteRow(profile profileRow) error {
 		if err = r.Rewrite(profile.row.StacktracePartitionID(), s.stacktraces); err != nil {
 			return
 		}
-		s.numSamples += uint64(len(values))
+		rewrittenSamples += uint64(len(values))
 		for i, v := range values {
 			// FIXME: the original order is not preserved, which will affect encoding.
 			values[i] = parquet.Int64Value(int64(s.stacktraces[i])).Level(v.RepetitionLevel(), v.DefinitionLevel(), v.Column())
 		}
 	})
 	if err != nil {
+		return rewrittenSamples, err
+	}
+	return rewrittenSamples, nil
+}
+
+func (s *symbolsCompactor) Close() error {
+	if s.flushed {
+		return nil
+	}
+	if err := s.w.Flush(); err != nil {
 		return err
 	}
+	s.flushed = true
 	return nil
 }
 
-func (s *symbolsRewriter) Close() error {
-	return s.w.Flush()
-}
-
-func (s *symbolsRewriter) loadStacktracesID(values []parquet.Value) {
+func (s *symbolsCompactor) loadStacktracesID(values []parquet.Value) {
 	s.stacktraces = grow(s.stacktraces, len(values))
 	for i := range values {
 		s.stacktraces[i] = values[i].Uint32()