--- a/pkg/phlaredb/testdata/01HHYG6245NWHZWVP27V8WJRT7/meta.json
+++ b/pkg/phlaredb/testdata/01HHYG6245NWHZWVP27V8WJRT7/meta.json
@@ -1,0 +1,86 @@
+{
+	"ulid": "01HHYG6245NWHZWVP27V8WJRT7",
+	"minTime": 1702904400.005,
+	"maxTime": 1702906199.998,
+	"stats": {
+		"numSamples": 1477784,
+		"numSeries": 362,
+		"numProfiles": 10000
+	},
+	"files": [
+		{
+			"relPath": "index.tsdb",
+			"sizeBytes": 163444,
+			"tsdb": {
+				"numSeries": 362
+			}
+		},
+		{
+			"relPath": "profiles.parquet",
+			"sizeBytes": 8169518,
+			"parquet": {
+				"numRowGroups": 1,
+				"numRows": 10000
+			}
+		},
+		{
+			"relPath": "symbols/functions.parquet",
+			"sizeBytes": 763574,
+			"parquet": {
+				"numRowGroups": 1,
+				"numRows": 19063
+			}
+		},
+		{
+			"relPath": "symbols/index.symdb",
+			"sizeBytes": 2140
+		},
+		{
+			"relPath": "symbols/locations.parquet",
+			"sizeBytes": 1407781,
+			"parquet": {
+				"numRowGroups": 1,
+				"numRows": 32293
+			}
+		},
+		{
+			"relPath": "symbols/mappings.parquet",
+			"sizeBytes": 2478,
+			"parquet": {
+				"numRowGroups": 1,
+				"numRows": 15
+			}
+		},
+		{
+			"relPath": "symbols/stacktraces.symdb",
+			"sizeBytes": 949513
+		},
+		{
+			"relPath": "symbols/strings.parquet",
+			"sizeBytes": 1008317,
+			"parquet": {
+				"numRowGroups": 1,
+				"numRows": 23085
+			}
+		}
+	],
+	"compaction": {
+		"level": 2,
+		"sources": [
+			"01HHYG628G63D5TBSK8BNX5Y07"
+		],
+		"parents": [
+			{
+				"ulid": "01HHYG628G63D5TBSK8BNX5Y07",
+				"minTime": 1702904400.005,
+				"maxTime": 1702906199.998
+			}
+		]
+	},
+	"version": 3,
+	"labels": {},
+	"source": "compactor",
+	"downsample": {
+		"resolution": 0
+	}
+}