… pkg/phlaredb/testdata/01HHYG6245NWHZWVP27V8WJRT7/index.tsdb
… pkg/phlaredb/testdata/01HHYG6245NWHZWVP27V8WJRT7/symbols/functions.parquet
… pkg/phlaredb/testdata/01HHYG6245NWHZWVP27V8WJRT7/symbols/index.symdb
… pkg/phlaredb/testdata/01HHYG6245NWHZWVP27V8WJRT7/symbols/locations.parquet
… pkg/phlaredb/testdata/01HHYG6245NWHZWVP27V8WJRT7/symbols/mappings.parquet
… pkg/phlaredb/testdata/01HHYG6245NWHZWVP27V8WJRT7/symbols/stacktraces.symdb
… pkg/phlaredb/testdata/01HHYG6245NWHZWVP27V8WJRT7/symbols/strings.parquet