This improve symbols compaction by using a single output symbols, this means we still merge all symbols, but we don't duplicate them anymore.

While this increase the amount of data, ultimately this is a temporary state when splitting blocks.
When merging blocks without splitting, symbols will still end up fully compacted as the original design.

EDIT: Change to symlink and we're now back same time.

```
❯ benchstat before.txt after.txt
name              old time/op    new time/op    delta
_CompactSplit-16     5.49s ±21%     4.09s ±17%     ~     (p=0.114 n=4+4)

name              old alloc/op   new alloc/op   delta
_CompactSplit-16    6.45GB ± 0%    3.32GB ± 0%  -48.54%  (p=0.029 n=4+4)

name              old allocs/op  new allocs/op  delta
_CompactSplit-16     2.89M ± 0%     0.43M ± 0%  -85.09%  (p=0.029 n=4+4)
```

<img width="2237" alt="image" src="https://github.com/grafana/pyroscope/assets/1053421/4e95f48e-467e-4b92-b98a-9ffdd17c3201">


Testing shortly in ops.