@kolesnikovae 2023-12-19
I very like the idea! The L2 vs L3 cost trade-off looks reasonable to me. I guess that now L3 requires significantly more memory than it used to be with split symbols, but it is probably still significantly less than the L2 peak. I'm wondering how we can make the merge phase cheaper (but still removing unreferenced symbols from the source blocks) – the first thing I would try is selective partitions loading: basically lazily load partitions _referenced_ in the block.