@cyriltovena 2023-12-19
> basically lazily load partitions _referenced_ in the block.

Yeah good thinking, I think also in the long run we might want to change how we split to account for symbols to avoid loading too much.